Refactor abstract base class for Invite*Fragments
